RavenDB-5177 Reduce the surface area of RavenThreadPool to make it easier to override
