+++ Raven.Database/Impl/BackgroundTaskExecuter/RavenThreadPool.cs
@@ -6,7 +6,6 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Threading;
 using Raven.Database.Config;
 
@@ -28,16 +27,14 @@ namespace Raven.Database.Impl.BackgroundTaskExecuter
         }
 
         public abstract ThreadsSummary GetThreadPoolStats();
-        public abstract ThreadTask[] GetRunningTasks();
-        public abstract ThreadTask[] GetAllWaitingTasks();
+        public abstract IEnumerable<object> GetRunningTaskDescriptions();
+        public abstract IEnumerable<object> GetAllWaitingTaskDescriptions();
         public abstract RavenThreadPool Start();
         public abstract void DrainThePendingTasks();
 
-        public abstract void ExecuteBatch<T>(IList<T> src, Action<T> action, string description = null,
-            bool allowPartialBatchResumption = false, int completedMultiplier = 2, int freeThreadsMultiplier = 2, int maxWaitMultiplier = 1);
+        public abstract void ExecuteBatch<T>(IList<T> src, Action<T> action, string description = null, bool allowPartialBatchResumption = false, int completedMultiplier = 2, int freeThreadsMultiplier = 2, int maxWaitMultiplier = 1);
 
-        public abstract void ExecuteBatch<T>(IList<T> src, Action<IEnumerator<T>> action,
-            int pageSize = DefaultPageSize, string description = null);
+        public abstract void ExecuteBatch<T>(IList<T> src, Action<IEnumerator<T>> action, int pageSize = DefaultPageSize, string description = null);
 
         protected virtual void Dispose(bool disposing)
         {
@@ -51,30 +48,6 @@ namespace Raven.Database.Impl.BackgroundTaskExecuter
             public int FreeThreadsAmount { get; set; }
             public int ConcurrentEventsCount { get; set; }
             public int ConcurrentWorkingThreadsAmount { get; set; }
-        }
-
-        public class ThreadTask
-        {
-            public Action Action;
-            public BatchStatistics BatchStats;
-            public object Description;
-            public CountdownEvent DoneEvent;
-            public Stopwatch Duration;
-            public bool EarlyBreak;
-            public DateTime QueuedAt;
-        }
-
-        public class BatchStatistics
-        {
-            public enum BatchType
-            {
-                Atomic,
-                Range
-            }
-
-            public int Completed;
-            public int Total;
-            public BatchType Type;
         }
     }
 }
